Restrict printing without completing registration Code clean up Language translations

+++ database/migrations/2016_08_19_172025_create_plot_reservation_table.php
@@ -35,7 +35,7 @@ class CreatePlotReservationTable extends Migration
 
             $table->timestamp('created_at');
 
-            $table->integer('status')->unsigned()->length(1)->defalt(0);
+            $table->tinyInteger('status')->unsigned()->length(1)->defalt(0);
 
             $table->primary(['area_id','areas_type_id','plot_id','block_id']);
         });
